move utils.random to test folder
