+++ lib/utils.js
@@ -237,16 +237,6 @@ exports.options = function(defaults, options) {
 };
 
 /*!
- * Generates a random string
- *
- * @api private
- */
-
-exports.random = function() {
-  return Math.random().toString().substr(3);
-};
-
-/*!
  * Merges `from` into `to` without overwriting existing properties.
  *
  * @param {Object} to
