+++ test/model.aggregate.test.js
@@ -9,7 +9,7 @@ const start = require('./common');
 
 const Aggregate = require('../lib/aggregate');
 const assert = require('assert');
-const random = require('../lib/utils').random;
+const random = require('./util').random;
 
 const mongoose = start.mongoose;
 const Schema = mongoose.Schema;
