+++ test/model.findOneAndUpdate.test.js
@@ -9,7 +9,7 @@ const start = require('./common');
 const CastError = require('../lib/error/cast');
 const assert = require('assert');
 const mongoose = start.mongoose;
-const random = require('../lib/utils').random;
+const random = require('./util').random;
 const Utils = require('../lib/utils');
 const Schema = mongoose.Schema;
 const ObjectId = Schema.Types.ObjectId;
