+++ test/model.test.js
@@ -7,7 +7,7 @@
 const start = require('./common');
 
 const assert = require('assert');
-const random = require('../lib/utils').random;
+const random = require('./util').random;
 const util = require('./util');
 
 const mongoose = start.mongoose;
