+++ test/schema.validation.test.js
@@ -8,7 +8,7 @@ const start = require('./common');
 
 const Promise = require('bluebird');
 const assert = require('assert');
-const random = require('../lib/utils').random;
+const random = require('./util').random;
 
 const mongoose = start.mongoose;
 const Schema = mongoose.Schema;
