+++ test/shard.test.js
@@ -4,7 +4,7 @@ const start = require('./common');
 
 const assert = require('assert');
 const chalk = require('chalk');
-const random = require('../lib/utils').random;
+const random = require('./util').random;
 
 const mongoose = start.mongoose;
 const Schema = mongoose.Schema;
